requirements.txt
tensorflow==2.12.0
fastapi==0.96.0
uvicorn==0.22.0
python-multipart==0.0.6
pillow==9.5.0
matplotlib==3.7.1
numpy==1.23.5
opencv-python==4.7.0.72